Updates application icons and fixes path resolution
Updates the application icons to use the correct file extensions for different platforms.

Addresses an issue where the application could not resolve the correct path for the protobuf definition in production environments, especially when packaged as an Electron app. It does this by adding a type assertion for the Electron's `resourcesPath` property.